requirements.txt
discord.py
asyncio
cloudflare
requests
PyMySQL
peewee